Fix RiscType enum cast issue: use uint64_t instead of uint32_t

+++ device/types/risc_type.cpp
@@ -119,9 +119,9 @@ std::string RiscTypeToString(RiscType value) {
 }
 
 RiscType invert_selected_options(RiscType selected) {
-    uint32_t selected_bits = static_cast<uint32_t>(selected);
-    uint32_t inverted =
-        (~selected_bits) & static_cast<uint32_t>(RiscType::ALL | RiscType::ALL_TENSIX | RiscType::ALL_NEO);
+    uint64_t selected_bits = static_cast<uint64_t>(selected);
+    uint64_t inverted =
+        (~selected_bits) & static_cast<uint64_t>(RiscType::ALL | RiscType::ALL_TENSIX | RiscType::ALL_NEO);
     return static_cast<RiscType>(inverted);
 }
